requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.1
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
defusedxml==0.7.1
deprecation==2.1.0
diff-match-patch==20230430
Django==4.2.4
django-admin-interface==0.26.0
django-colorfield==0.9.0
django-cors-headers==4.2.0
django-countries==7.5.1
django-debug-toolbar==4.2.0
django-filter==23.2
django-import-export==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
flower==2.0.1
humanize==4.8.0
img2pdf==0.4.4
kombu==5.3.1
lxml==4.9.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pikepdf==8.4.0
Pillow==10.0.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
text-unidecode==1.3
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
